refactor: split up request module into more files
- delete unlinked files
- add optional count method for Requests
