+++ src/db/requests/mod.rs
@@ -1,16 +1,21 @@
 use super::*;
 
 pub mod prelude {
+    pub use super::Request;
+    pub use super::db_request::DbRequest;
     pub use super::handle_requests::handle_requests;
-    pub use super::{DbRequest, Request, RequestType};
+    pub use super::req_type::RequestType;
 }
 
+mod db_request;
 mod handle_requests;
+mod req_type;
 
 pub trait Request<'e> {
     type Output;
     type Connection: SqliteExecutor<'e>;
 
+    // required methods
     fn check_id(&self, req_type: RequestType) -> Result<i64, RequestError>;
     fn get(
         &self,
@@ -32,81 +37,9 @@ pub trait Request<'e> {
         &self,
         conn: Self::Connection,
     ) -> impl Future<Output = Result<u64>>;
-}
 
-#[derive(Debug, Default, PartialEq)]
-pub struct DbRequest {
-    pub req_type: RequestType,
-    pub payload: DbPayload,
-}
-
-impl DbRequest {
-    /// Returns an array of DbRequests related to fetching all table data
-    /// with offsets of 0.
-    pub fn init() -> Vec<Self> {
-        [
-            DbPayload::ItemParams(ItemParams::default()),
-            DbPayload::UserParams(UserParams::default()),
-            DbPayload::ReceiptParams(JoinedReceiptParams::default()),
-        ]
-        .into_iter()
-        .map(|payload| {
-            let mut req = Self::new();
-            req.with_req_type(RequestType::GetAll).with_payload(payload);
-            req
-        })
-        .collect()
-    }
-    // TODO: make refresh offsets and limits configurable
-    /// Returns a pre-built DbRequest for refetching StoreTotals and table data.
-    /// This is Vec with a Refresh Requests and the three init requests.
-    pub fn refresh() -> Vec<Self> {
-        let mut refresh = Self::new();
-        refresh.with_req_type(RequestType::Refresh);
-
-        let mut requests = Vec::with_capacity(4);
-        requests.push(refresh);
-        requests.append(&mut Self::init());
-
-        requests
-    }
-    pub fn new() -> Self {
-        Self::default()
-    }
-    pub fn with_req_type(&mut self, req_type: RequestType) -> &mut Self {
-        self.req_type = req_type;
-        self
-    }
-    pub fn with_payload(&mut self, payload: DbPayload) -> &mut Self {
-        self.payload = payload;
-        self
-    }
-}
-
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub enum RequestType {
-    #[default]
-    None,
-    GetAll,
-    Get,
-    Post,
-    Update,
-    Delete,
-    Reset,
-    Refresh,
-}
-
-impl Display for RequestType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::None => write!(f, "None"),
-            Self::GetAll => write!(f, "Get all"),
-            Self::Get => write!(f, "Get"),
-            Self::Post => write!(f, "Post"),
-            Self::Update => write!(f, "Update"),
-            Self::Delete => write!(f, "Delete"),
-            Self::Reset => write!(f, "Reset"),
-            Self::Refresh => write!(f, "Refresh"),
-        }
+    // optional methods
+    fn count() -> impl Future<Output = Result<u64>> {
+        async { Ok(0) }
     }
 }
